Renamet content til resources
Lest oppgaven på nytt og innså at den hadde fail navn lol
